Remove unnecessary field
