+++ subprojects/launcher/src/main/java/org/gradle/tooling/internal/provider/events/DefaultTestSkippedResult.java
@@ -20,6 +20,6 @@ import org.gradle.tooling.internal.protocol.events.InternalTestSkippedResult;
 
 public class DefaultTestSkippedResult extends AbstractTestResult implements InternalTestSkippedResult {
     public DefaultTestSkippedResult(long startTime, long endTime) {
-        super(startTime, endTime, "skipped", "TODO not needed", "TODO not needed");
+        super(startTime, endTime, "skipped");
     }
 }
