+++ subprojects/launcher/src/main/java/org/gradle/tooling/internal/provider/events/DefaultTestSuccessResult.java
@@ -20,6 +20,6 @@ import org.gradle.tooling.internal.protocol.events.InternalTestSuccessResult;
 
 public class DefaultTestSuccessResult extends AbstractTestResult implements InternalTestSuccessResult {
     public DefaultTestSuccessResult(long startTime, long endTime) {
-        super(startTime, endTime, "succeeded", "TODO: remove", "TODO: remove");
+        super(startTime, endTime, "succeeded");
     }
 }
